Maze Creator
This function returns a 2D array to define what is or isn't a wall on the map. Ones are wall. Zeros are floor. Once we have wall objects we can iterate through the array to place them down in the proper locations.

Guaranteed to have every floor tile reachable from any other and absolutely no dead ends.